1-Basic-Web-App-Creation-And-Configuration/src/requirements.txt
flask==3.0.0
gunicorn==23.0.0
psycopg2==2.9.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
